moved workflowOutputParameter
